wallet: refactor wallet.go by moving legacy methods to deprecated.go
This commit refactors wallet.go to be a lean central coordination point,
containing only the lifecycle (Open/Create), core configuration, and
modern/stable API methods.

All legacy methods, including RPC-specific logic, deprecated helpers
(like InitAccounts, AddScopeManager, DeriveFromKeyPath), and methods
replaced by new Managers, have been moved to deprecated.go.

The walletDeprecated struct is used to encapsulate legacy state.
